+++ src/Float_ResizeEnds.cpp
@@ -2,8 +2,8 @@
 
 void InfiniteArithmetic::Float::ResizeEnds(Float &num1, Float &num2)
 {
-	size_t pp1 = num1.PointPosition;
-	size_t pp2 = num2.PointPosition;
+	int64_t pp1 = num1.PointPosition;
+	int64_t pp2 = num2.PointPosition;
 
 	if(pp1 > pp2)
 	{
